Feat: Base da lógica de Método do Cotovelo

+++ tests/include/sse.h
@@ -1,6 +1,6 @@
 #ifndef SSE_H
 #define SSE_H
 
-double calculateSSE(const unsigned char *const data, unsigned char *const clusters, const unsigned char *const centroids, unsigned size);
+unsigned long long calculateSSE(const unsigned char *const data, unsigned char *const clusters, const unsigned char *const centroids, unsigned size);
 
 #endif
