+++ tests/src/sse.c
@@ -1,6 +1,6 @@
 #include "sse.h"
 
-double calculateSSE(const unsigned char *const data, unsigned char *const clusters, const unsigned char *const centroids, unsigned size) {
+unsigned long long calculateSSE(const unsigned char *const data, unsigned char *const clusters, const unsigned char *const centroids, unsigned size) {
     unsigned long long sse = 0;
     for (unsigned i = 0; i < size; i++) {
         int error = (int)data[i] - (int)centroids[clusters[i]];
